(add) Player 1 and 2 (each self contained) + round timer + player health bar + round state + tiebreaker

+++ Assets/Scripts/Weapons/WeaponData.cs
@@ -14,7 +14,7 @@ public class WeaponData : ScriptableObject
     public int minWordLength;
     public int maxWordLength;
 
-    public int damagePerShot;
+    public float damagePerShot;
     public float fireRateModifier;
 
     public SpecialEffect specialEffect;
